% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_data}
\alias{create_data}
\title{Create a 2D Synthetic Dataset With Class Labels}
\usage{
create_data(
  coord_filepath = NULL,
  coord = NULL,
  n_samples = 10000,
  type = "uniform",
  seed = 1835,
  gap = 0
)
}
\arguments{
\item{coord_filepath}{A character string specifying a file path to a JSON
file containing a list/array of coordinate pairs. Optional if \code{coord} is
supplied.}

\item{coord}{A matrix or data frame with two columns (x and y) specifying
threshold coordinates. Optional if \code{coord_filepath} is supplied.}

\item{n_samples}{Integer; number of synthetic 2D points to generate.}

\item{type}{String; How should the points be generated. Can be either 'uniform' or 'grid'}

\item{seed}{Integer; Random seed to be used for generating dataset}

\item{gap}{Numeric; Value close to 0 defining gap at the boundary, less than 0.2}
}
\value{
A data frame with three columns:
\describe{
\item{x}{Random uniform x-coordinate}
\item{y}{Random uniform y-coordinate}
\item{class}{Binary class label ("Above" or "Below")}
}
}
\description{
Generates a synthetic two-dimensional dataset using random uniform values.
Each point is assigned a class label based on whether it lies above all
coordinate thresholds provided. Coordinates can be supplied directly as a
matrix/data frame or indirectly via a JSON file containing an array of
coordinate pairs.
}
\details{
A point is labelled \code{class = 1} if it is above \strong{all} coordinate
thresholds such that:
\deqn{x > coord[,1] \; \text{and} \; y > coord[,2]}

Otherwise, the point is labelled \code{0}.
}
\examples{

coords <- matrix(c(0.2, 0.3,
                   0.4, 0.5,
                   0.6, 0.7), ncol = 2, byrow = TRUE)

df <- create_data(coord = coords, n_samples = 500, seed = 717)
head(df)


}
