% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sillysplinesOutput}
\alias{sillysplinesOutput}
\alias{renderSillysplines}
\title{Shiny bindings for sillysplines}
\usage{
sillysplinesOutput(outputId, width = "100\%", height = "400px")

renderSillysplines(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates an HTML widget (or a
\href{https://rstudio.github.io/promises/}{promise} of an HTML widget).}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using sillysplines within Shiny
applications and interactive Rmd documents.
}
\examples{
\dontshow{if (rlang::is_installed(c("shiny"))) withAutoprint(\{ # examplesIf}
library(shiny)
app <- shinyApp(
  ui = fluidPage(sillysplinesOutput('splines')),
  server = function(input, output) {
    splines_widget <- sillysplines()
    output$splines= renderSillysplines(splines_widget)
  }
)

\donttest{if (interactive()) app}
\dontshow{\}) # examplesIf}
}
