% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sillysplines.R
\name{sillysplines}
\alias{sillysplines}
\title{Create a box to draw silly spline}
\usage{
sillysplines(width = 640, height = 640, elementId = "app")
}
\arguments{
\item{width}{width of container in pixels}

\item{height}{height of container in pixels}

\item{elementId}{The id to be used for the container element}
}
\description{
This creates a simple htmlwidget to be used
}
