% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sillysplines.R
\name{sillysplines}
\alias{sillysplines}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
sillysplines(message, width = NULL, height = NULL, elementId = NULL)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
